+++ checkpoint/orbax/checkpoint/experimental/v1/__init__.py
@@ -25,6 +25,7 @@ from orbax.checkpoint.experimental.v1 import handlers
 from orbax.checkpoint.experimental.v1 import training
 from orbax.checkpoint.experimental.v1 import path
 from orbax.checkpoint.experimental.v1 import errors
+from orbax.checkpoint.experimental.v1 import serialization
 from orbax.checkpoint.experimental.v1._src.context import options
 from orbax.checkpoint.experimental.v1._src.context.context import (
     Context,