+++ checkpoint/CHANGELOG.md
@@ -22,6 +22,7 @@ centralize definitions.
 
 - Add support for loading SafeTensors checkpoints
 - #v1 Add `is_orbax_checkpoint()` method for validation checks
+- #v1 Add custom LeafHandler docs.
 
 ### Changed